orders/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
structlog==23.2.0
httpx==0.25.2
pika==1.3.2
sqladmin==0.15.0
python-jose[cryptography]==3.3.0
requests==2.31.0
pytest==7.4.3
email-validator==2.1.0
itsdangerous==2.1.2
PyJWT==2.8.0 